feat(kubeconform): allow using custom resource schemas
Allowing to pass custom -schema-location arguments enables us to
validate custom Kubernetes resources that are not included
in the default set that kubeconform validates.

We need to parse arguments because one argument is used to
select the overlay directories to process with kustomize, while
we also want to pass the list of schema locations to kubeconform.

Use arguments with single -, that form is documented in kubeconform
docs.

Co-authored-by: Alex Viscreanu <alexviscreanu@gmail.com>